get all reviews associated with a single product

+++ controllers/reviewController.js
@@ -63,15 +63,28 @@ const updateReview = async (req, res) => {
 const deleteReview = async (req, res) => {
   const { id: reviewId } = req.params;
   const review = await Review.findOne({ _id: reviewId });
+  if (!review) {
+    throw new CustomError.NotFoundError(`No review with id ${reviewId}`);
+  }
   checkPermissions(req.user, review.user);
   await review.deleteOne();
   res.status(StatusCodes.OK).json({ msg: 'Review successfully removed' });
 };
 
+const getSingleProductReviews = async (req, res) => {
+  const { id: productId } = req.params;
+  const product = await Product.findOne({ _id: productId });
+  if (!product) {
+    throw new CustomError.NotFoundError(`No product with id ${productId}`);
+  }
+  const reviews = await Review.find({ product: productId });
+  res.status(StatusCodes.OK).json({ count: reviews.length, reviews });
+};
 module.exports = {
   createReview,
   getAllReviews,
   getSingleReview,
   updateReview,
   deleteReview,
+  getSingleProductReviews,
 };
